Survey row and component properties
