+++ tests/test_models.py
@@ -3,70 +3,21 @@ import sys
 
 import pytest
 
-from django.core.management import call_command
-
-from countries_plus.models import Country
-from pldp.models import Agency, Survey, SurveyRow, SurveyComponent, StudyArea, \
-    Study, Location, LocationLine
+from pldp.models import Survey, SurveyRow, SurveyComponent
 from pldp.forms import GENDER_BASIC_CHOICES, AGE_BASIC_CHOICES, \
     AGE_DETAILED_CHOICES, AGE_COMPLEX_CHOICES
 
-@pytest.fixture
+
 @pytest.mark.django_db
-def countries(db):
-    call_command('update_countries_plus')
+def test_survey_row(survey, survey_row, survey_component):
+    row = SurveyRow.objects.get(survey_id=survey.id)
+
+    assert survey.row == row
 
 
-@pytest.fixture
-def country(countries):
-    return Country.objects.order_by('?').first()
-
-
-@pytest.fixture
-def agency():
-    agency = Agency.objects.create(name='Test Agency',
-                                   email='test@testagency.com')
-    return agency
-
-
-@pytest.fixture
-def location_line():
-    linestring = 'LINESTRING(-101.744384 39.32155, -101.552124 39.330048, -101.403808 39.330048, -101.332397 39.364032)'
-    return LocationLine.objects.create(geometry=linestring,
-                                       date_measured='2018-11-11')
-
-
-@pytest.fixture
-def location(agency, country, location_line):
-    location = Location.objects.create(agency=agency,
-                                       country=country,
-                                       primary_name='Test location',
-                                       line=location_line)
-    return location
-
-
-@pytest.fixture
-def study_area():
-    polygon = 'POLYGON((-101.744384 39.32155, -101.552124 39.330048, -101.403808 39.330048, -101.332397 39.364032, -101.744384 39.32155))'
-
-    return StudyArea.objects.create(name='Test study area',
-                                    area=polygon)
-
-
-@pytest.fixture
-def study(study_area, agency):
-    study = Study.objects.create(title='Test study',
-                                 agency=agency)
-    study.areas.add(study_area)
-    return study
-
-
-@pytest.fixture
-def survey(location, study):
-    return Survey.objects.create(study=study,
-                                 location=location,
-                                 form_id=1,
-                                 method='analog')
+@pytest.mark.django_db
+def test_survey_components(survey, survey_row, survey_component):
+    assert str(survey.components.first().label) == 'Test label'
 
 
 @pytest.mark.django_db
